Response to more code review comments
